17 - DatabaseManager deleted
